Created HoloTable.java
One variable of each sensor was made as an example for when they are actually added in. The variables are also assigned their hypothetical port value. Hashmap code has been commented out incase we go back to it. Places for the camera and color sensors are included, but the variables themselves were not created(Left camera for Brendan and Jack)